feature(todo task list): update component
