+++ frontend/src/components/TodoTaskList/TodoTaskList.styled.tsx
@@ -1,4 +1,9 @@
+import { List } from 'antd';
 import styled from 'styled-components';
+
+export const StyledList = styled(List)`
+  background-color: #fff;
+`;
 
 export const StyledTodoTaskListFooter = styled.footer`
   font-weight: 600;
