--- frontend/src/components/TodoTaskList/TodoTaskListFooter/TodoTaskListFooterProps.ts
@@ -1,4 +1,0 @@
-export type TodoTaskListFooterProps = {
-  completedTaskVisible: boolean;
-  toggleTheVisibilityOfDoneTasks: () => void;
-};
