--- frontend/src/components/TodoTaskList/TodoTaskListFooter/index.tsx
@@ -1,15 +1,0 @@
-import { TodoTaskListFooterProps } from './TodoTaskListFooterProps';
-import { StyledTodoTaskListFooter } from './index.styled';
-
-export const TodoTaskListFooter = ({
-  completedTaskVisible,
-  toggleTheVisibilityOfDoneTasks,
-}: TodoTaskListFooterProps) => {
-  return (
-    <StyledTodoTaskListFooter>
-      <button onClick={toggleTheVisibilityOfDoneTasks}>
-        {completedTaskVisible ? 'Hide complete tasks' : 'Show complete tasks'}
-      </button>
-    </StyledTodoTaskListFooter>
-  );
-};
