--- frontend/src/components/TodoTaskList/index.styled.tsx
@@ -1,6 +1,0 @@
-import { List } from 'antd';
-import styled from 'styled-components';
-
-export const StyledList = styled(List)`
-  background-color: #fff;
-`;
