+++ frontend/src/views/TodoList/TodoList.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { usePageTitle } from 'providers/PageTitleProvider';
-import { TodoTaskList } from 'components/TodoTaskList';
-import { todoTaskListMock } from 'mock-data/TodoTaskList/TodotaskListMock';
+import { TodoTaskList } from 'components/TodoTaskList/TodoTaskList';
+import { todoTaskListMock } from 'mock-data/TodoTaskList/todotaskListMock';
 
 export const TodoList = () => {
   const { setTitle } = usePageTitle();
